Extracting logic into search handler
