+++ trustar2/handlers/indicators.py
@@ -1,7 +1,8 @@
 from __future__ import unicode_literals
 
 from trustar2.query import Query
-from trustar2.handlers.base_handler import BaseHandler
+from trustar2.handlers.base_handler import BaseHandler # remove
+from trustar2.handlers.search_handler import SearchHandler
 from trustar2.base import fluent, Methods, get_timestamp
 from trustar2.handlers.tags import TagIndicator
 from trustar2.trustar_enums import ObservableTypes, SortColumns, SortOrder, AttributeTypes
@@ -10,7 +11,7 @@ from trustar2.models import Entity
 
 
 @fluent
-class SearchIndicator(BaseHandler):
+class SearchIndicator(SearchHandler):
     url = "/indicators"
 
     def __init__(self, config=None):
@@ -23,20 +24,20 @@ class SearchIndicator(BaseHandler):
         return self.config.request_details.get("api_endpoint") + self.url
 
         
-    @staticmethod
-    def _get_value(arg, enum):
-        if isinstance(arg, enum):
-            return arg.value
+    # @staticmethod
+    # def _get_value(arg, enum):
+    #     if isinstance(arg, enum):
+    #         return arg.value
 
-        if isinstance(arg, str) and arg in enum.members():
-            return arg
+    #     if isinstance(arg, str) and arg in enum.members():
+    #         return arg
 
-        if isinstance(arg, type("")) and arg in enum.members():
-            return arg  # For py2 and py3 compatibility
+    #     if isinstance(arg, type("")) and arg in enum.members():
+    #         return arg  # For py2 and py3 compatibility
 
-        raise AttributeError(
-            "Possible value types are: {}".format(list(enum.members()))
-        )
+    #     raise AttributeError(
+    #         "Possible value types are: {}".format(list(enum.members()))
+    #     )
 
     @staticmethod
     def _get_entity(entity, entity_type):
@@ -59,30 +60,30 @@ class SearchIndicator(BaseHandler):
         )
 
 
-    def set_query_term(self, query):
-        self.set_payload_param("queryTerm", query)
+    # def set_query_term(self, query):
+    #     self.set_payload_param("queryTerm", query)
 
 
-    def set_page_size(self, page_size):
-        self.set_query_param("pageSize", page_size)
+    # def set_page_size(self, page_size):
+    #     self.set_query_param("pageSize", page_size)
 
 
-    def set_from(self, from_date):
-        if not isinstance(from_date, int):
-            from_date = get_timestamp(from_date)
-        self.set_payload_param("from", from_date)
+    # def set_from(self, from_date):
+    #     if not isinstance(from_date, int):
+    #         from_date = get_timestamp(from_date)
+    #     self.set_payload_param("from", from_date)
 
 
-    def set_to(self, to_date):
-        if not isinstance(to_date, int):
-            to_date = get_timestamp(to_date)
+    # def set_to(self, to_date):
+    #     if not isinstance(to_date, int):
+    #         to_date = get_timestamp(to_date)
 
-        self.set_payload_param("to", to_date)
+    #     self.set_payload_param("to", to_date)
 
 
-    def set_sort_column(self, column):
-        column = self._get_value(column, SortColumns)
-        self.set_payload_param("sortColumn", column)
+    # def set_sort_column(self, column):
+    #     column = self._get_value(column, SortColumns)
+    #     self.set_payload_param("sortColumn", column)
 
 
     def set_sort_order(self, order):
@@ -96,22 +97,22 @@ class SearchIndicator(BaseHandler):
         self.set_payload_param("priorityScores", scores)
 
 
-    def set_enclave_ids(self, enclave_guids):
-        if not isinstance(enclave_guids, list):
-            enclave_guids = [enclave_guids]
-        self.set_payload_param("enclaveGuids", enclave_guids)
+    # def set_enclave_ids(self, enclave_guids):
+    #     if not isinstance(enclave_guids, list):
+    #         enclave_guids = [enclave_guids]
+    #     self.set_payload_param("enclaveGuids", enclave_guids)
 
 
-    def set_included_tags(self, tags):
-        if not isinstance(tags, list):
-            tags = [tags]
-        self.set_payload_param("includedTags", tags)
+    # def set_included_tags(self, tags):
+    #     if not isinstance(tags, list):
+    #         tags = [tags]
+    #     self.set_payload_param("includedTags", tags)
 
 
-    def set_excluded_tags(self, tags):
-        if not isinstance(tags, list):
-            tags = [tags]
-        self.set_payload_param("excludedTags", tags)
+    # def set_excluded_tags(self, tags):
+    #     if not isinstance(tags, list):
+    #         tags = [tags]
+    #     self.set_payload_param("excludedTags", tags)
 
 
     def set_observable_types(self, types):
@@ -154,6 +155,7 @@ class SearchIndicator(BaseHandler):
             .set_params(self.payload_params)
             .set_query_string({"pageSize": self.query_params.get("pageSize", 25)})
         )
+
 
     def tags(self):
         return TagIndicator(self.config)
