+++ trustar2/handlers/submission.py
@@ -2,13 +2,14 @@ from __future__ import unicode_literals
 
 from trustar2.query import Query
 from trustar2.trustar_enums import MaxValues
-from trustar2.handlers.base_handler import BaseHandler
+from trustar2.handlers.base_handler import BaseHandler # remove
+from trustar2.handlers.search_handler import SearchHandler
 from trustar2.handlers.tags import TagSubmission
 from trustar2.base import fluent, Methods, get_timestamp
 
 
 @fluent
-class Submission(BaseHandler):
+class Submission(SearchHandler):
 
     SUBMISSION_MANDATORY_FIELDS = ("title", "content", "enclaveGuid")
     _path = "/submissions"
@@ -185,48 +186,48 @@ class Submission(BaseHandler):
         return query_params
 
 
-    def set_page_size(self, page_size):
-        self.set_query_param("pageSize", page_size)
+    # def set_page_size(self, page_size):
+    #     self.set_query_param("pageSize", page_size)
 
     
-    def set_query_term(self, query):
-        self.set_payload_param("queryTerm", query)
+    # def set_query_term(self, query):
+    #     self.set_payload_param("queryTerm", query)
 
 
-    def set_from(self, from_date):
-        if not isinstance(from_date, int):
-            from_date = get_timestamp(from_date)
-        self.set_payload_param("from", from_date)
+    # def set_from(self, from_date):
+    #     if not isinstance(from_date, int):
+    #         from_date = get_timestamp(from_date)
+    #     self.set_payload_param("from", from_date)
 
 
-    def set_to(self, to_date):
-        if not isinstance(to_date, int):
-            to_date = get_timestamp(to_date)
+    # def set_to(self, to_date):
+    #     if not isinstance(to_date, int):
+    #         to_date = get_timestamp(to_date)
 
-        self.set_payload_param("to", to_date)
+    #     self.set_payload_param("to", to_date)
 
 
-    def set_enclave_ids(self, enclave_guids):
-        if not isinstance(enclave_guids, list):
-            enclave_guids = [enclave_guids]
-        self.set_payload_param("enclaveGuids", enclave_guids)
+    # def set_enclave_ids(self, enclave_guids):
+    #     if not isinstance(enclave_guids, list):
+    #         enclave_guids = [enclave_guids]
+    #     self.set_payload_param("enclaveGuids", enclave_guids)
 
     
-    def set_included_tags(self, tags):
-        if not isinstance(tags, list):
-            tags = [tags]
-        self.set_payload_param("includedTags", tags)
+    # def set_included_tags(self, tags):
+    #     if not isinstance(tags, list):
+    #         tags = [tags]
+    #     self.set_payload_param("includedTags", tags)
 
 
-    def set_excluded_tags(self, tags):
-        if not isinstance(tags, list):
-            tags = [tags]
-        self.set_payload_param("excludedTags", tags)
+    # def set_excluded_tags(self, tags):
+    #     if not isinstance(tags, list):
+    #         tags = [tags]
+    #     self.set_payload_param("excludedTags", tags)
         
 
-    def set_sort_column(self, column):
-        column = self._get_value(column, SortColumns)
-        self.set_payload_param("sortColumn", column)
+    # def set_sort_column(self, column):
+    #     column = self._get_value(column, SortColumns)
+    #     self.set_payload_param("sortColumn", column)
 
 
     def should_use_external_id(self):
